[FIX] torch explicit dtype + remove Deprecated Variable

+++ leaspy/io/data/data.py
@@ -132,7 +132,7 @@ class Data:
             for j, age in enumerate(ages):
                 indices.append(indiv.idx)
                 timepoints[iteration] = age
-                arr[iteration] = torch.tensor(indiv.observations[j])
+                arr[iteration] = torch.tensor(indiv.observations[j], dtype=torch.float32)
 
                 iteration += 1
